ZO-3662: Move connection utility to zeit.simplecast
