+++ core/src/zeit/simplecast/configure.zcml
@@ -3,4 +3,7 @@
   xmlns:zcml="http://namespaces.zope.org/zcml"
   xmlns:grok="http://namespaces.zope.org/grok"
   i18n_domain="zeit.cms">
+
+  <grok:grok package="." />
+
 </configure>
