+++ core/src/zeit/simplecast/connection.py
@@ -2,17 +2,17 @@ import logging
 import requests
 import grokcore.component as grok
 import zope.app.appsetup.product
-import zeit.content.audio.interfaces
+import zeit.simplecast.interfaces
 
 log = logging.getLogger(__name__)
 
 
-@grok.implementer(zeit.content.audio.interfaces.ISimplecast)
+@grok.implementer(zeit.simplecast.interfaces.ISimplecast)
 class Simplecast(grok.GlobalUtility):
 
     def __init__(self):
         config = zope.app.appsetup.product.getProductConfiguration(
-            'zeit.content.audio')
+            'zeit.simplecast')
         self.api_url = config['simplecast-url']
         self.api_token = f"Bearer {config['simplecast-token']}"
