+++ core/src/zeit/simplecast/ftesting.zcml
@@ -7,4 +7,7 @@
   <include package="zeit.content.audio" />
   <include package="zeit.content.audio.browser" />
 
+  <include package="zeit.simplecast" />
+  <include package="zeit.simplecast.json" />
+
 </configure>
