+++ core/src/zeit/simplecast/testing.py
@@ -1,9 +1,19 @@
 import zeit.cms.testing
 
 
+product_config = """\
+<product-config zeit.simplecast>
+  simplecast-url https://testapi.simplecast.com/
+  simplecast-token TkQvZUd2MHRnR0UybFhsgTfs
+</product-config>
+"""
+
+CONFIG_LAYER = zeit.cms.testing.ProductConfigLayer(
+    product_config,
+    bases=(zeit.cms.testing.CONFIG_LAYER,))
 ZCML_LAYER = zeit.cms.testing.ZCMLLayer(
     'ftesting.zcml',
-    bases=(zeit.cms.testing.CONFIG_LAYER,))
+    bases=(CONFIG_LAYER,))
 ZOPE_LAYER = zeit.cms.testing.ZopeLayer(bases=(ZCML_LAYER,))
 WSGI_LAYER = zeit.cms.testing.WSGILayer(bases=(ZOPE_LAYER,))
